[UPD] contents Avg:starRate

+++ server/BED/src/main/java/main008/BED/contents/dto/ContentsDto.java
@@ -56,7 +56,7 @@ public class ContentsDto {
         private String thumbnail;
         private int likesCount;
         private Contents.Categories categories;
-        private float grade;
+        private double grade;
         private String tutorName;
         private String details;
         private String tutorDetail;
